backend/requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
celery>=5.3.0
stripe==6.0.0
djangorestframework
djangorestframework-simplejwt
django-filter
django-nested-admin
psycopg2-binary
gunicorn
redis
python-dotenv
django-cors-headers
whitenoise
pytest
pytest-django
pytest-cov
model-bakery
# Optional for API docs
drf-yasg
